Expand range of Digipot and Duration to 1024

+++ stm32f4-test-platformio/src/stm32f4xx_it.c
@@ -42,7 +42,7 @@ extern uint16_t state;
 
 extern uint8_t modeNumber;
 extern uint8_t lightNumber[3];
-extern uint8_t stepExper;
+extern uint16_t stepExper;
 extern uint8_t spiSent;
 /* USER CODE END 0 */
